pyproject.toml
[tool.poetry]
name = "superbenchmark"
version = "0.1.0"
description = ""
authors = ["paulmusquaro <136118097+paulmusquaro@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
pydantic = "^2.10.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^8.3.3"
httpx = "^0.27.2"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
